Update README, typo fixes, command line arg updates
Command line args now take hyphens ("-") instead of underscores ("_").

New flag no_init has been added. It won't run the init export when this
is set.

A small bug fix has been added for the working_dir option. It would be
automatically set to "." if an option was not specified and the program
would chdir to that by default (I don't exactly remember if this was the
fix or not).